+++ src/output/render/mod.rs
@@ -38,3 +38,6 @@ pub use self::users::Colours as UserColours;
 
 mod octal;
 // octal uses just one colour
+
+mod flags;
+// flags uses just one colour